feat: Fully migrate to multiple simulations
