+++ app/src/store.ts
@@ -3,7 +3,6 @@ import flowReducer from './reducers/flowReducer';
 import metadataReducer from './reducers/metadataReducer';
 import policyReducer from './reducers/policyReducer';
 import populationReducer from './reducers/populationReducer';
-import simulationReducer from './reducers/simulationReducer';
 import simulationsReducer from './reducers/simulationsReducer';
 
 export const store = configureStore({
@@ -11,10 +10,9 @@ export const store = configureStore({
     policy: policyReducer,
     flow: flowReducer,
     household: populationReducer,
-    simulation: simulationReducer,
+    simulations: simulationsReducer,
     population: populationReducer,
     metadata: metadataReducer,
-    simulations: simulationsReducer,
   },
 });
