+++ app/src/tests/frames/SimulationCreationFrame.test.tsx
@@ -3,7 +3,6 @@ import { render, screen, userEvent } from '@test-utils';
 import { Provider } from 'react-redux';
 import { configureStore } from '@reduxjs/toolkit';
 import SimulationCreationFrame from '@/frames/SimulationCreationFrame';
-import simulationReducer from '@/reducers/simulationReducer';
 import simulationsReducer from '@/reducers/simulationsReducer';
 import flowReducer from '@/reducers/flowReducer';
 import policyReducer from '@/reducers/policyReducer';
@@ -16,10 +15,9 @@ import {
 } from '@/tests/fixtures/frames/SimulationCreationFrame';
 
 // Mock actions to spy on
-import * as simulationActions from '@/reducers/simulationReducer';
 import * as simulationsActions from '@/reducers/simulationsReducer';
 
-describe('SimulationCreationFrame - Dual Dispatch', () => {
+describe('SimulationCreationFrame', () => {
   let store: any;
   let mockOnNavigate: ReturnType<typeof vi.fn>;
   let mockOnReturn: ReturnType<typeof vi.fn>;
@@ -31,7 +29,6 @@ describe('SimulationCreationFrame - Dual Dispatch', () => {
     // Create a fresh store for each test
     store = configureStore({
       reducer: {
-        simulation: simulationReducer,
         simulations: simulationsReducer,
         flow: flowReducer,
         policy: policyReducer,
@@ -59,7 +56,6 @@ describe('SimulationCreationFrame - Dual Dispatch', () => {
     };
     
     // Spy on the action creators
-    vi.spyOn(simulationActions, 'updateSimulationLabel');
     vi.spyOn(simulationsActions, 'updateSimulationLabel');
     vi.spyOn(simulationsActions, 'createSimulation');
   });
@@ -79,7 +75,7 @@ describe('SimulationCreationFrame - Dual Dispatch', () => {
     expect(Object.keys(state.simulations.entities)).toHaveLength(1);
   });
 
-  test('given user submits label then dispatches to both reducers', async () => {
+  test('given user submits label then dispatches to simulations reducer', async () => {
     // Given
     const user = userEvent.setup();
     render(
@@ -95,19 +91,15 @@ describe('SimulationCreationFrame - Dual Dispatch', () => {
     await user.type(input, TEST_SIMULATION_LABEL);
     await user.click(submitButton);
     
-    // Then - should dispatch to old reducer
-    expect(simulationActions.updateSimulationLabel).toHaveBeenCalledWith(TEST_SIMULATION_LABEL);
+    // Then - should dispatch to simulations reducer
+    expect(simulationsActions.updateSimulationLabel).toHaveBeenCalledWith({ label: TEST_SIMULATION_LABEL });
     
-    // And - should dispatch to new reducer
-    expect(simulationsActions.updateSimulationLabel).toHaveBeenCalledWith({
-      label: TEST_SIMULATION_LABEL,
-    });
     
     // And - should navigate to next
     expect(mockOnNavigate).toHaveBeenCalledWith('next');
   });
 
-  test('given user submits label then both reducer states are updated', async () => {
+  test('given user submits label then reducer state is updated', async () => {
     // Given
     const user = userEvent.setup();
     render(
@@ -123,11 +115,8 @@ describe('SimulationCreationFrame - Dual Dispatch', () => {
     await user.type(input, TEST_SIMULATION_LABEL);
     await user.click(submitButton);
     
-    // Then - check old reducer state
+    // Then - check reducer state
     const state = store.getState();
-    expect(state.simulation.label).toBe(TEST_SIMULATION_LABEL);
-    
-    // And - check new reducer state
     const activeId = state.simulations.activeId;
     expect(state.simulations.entities[activeId].label).toBe(TEST_SIMULATION_LABEL);
   });
@@ -153,7 +142,7 @@ describe('SimulationCreationFrame - Dual Dispatch', () => {
     expect(finalActiveId).toBe(initialActiveId);
   });
 
-  test('given empty label then still dispatches to both reducers', async () => {
+  test('given empty label then still dispatches to reducer', async () => {
     // Given
     const user = userEvent.setup();
     render(
@@ -167,10 +156,7 @@ describe('SimulationCreationFrame - Dual Dispatch', () => {
     // When - submit without entering a label
     await user.click(submitButton);
     
-    // Then - should dispatch empty string to old reducer
-    expect(simulationActions.updateSimulationLabel).toHaveBeenCalledWith('');
-    
-    // And - should dispatch empty string to new reducer
+    // Then - should dispatch empty string to reducer
     expect(simulationsActions.updateSimulationLabel).toHaveBeenCalledWith({
       label: '',
     });
